pynodegl: rework parameter init functions
This notably fixes setting _Node parameters as variable arguments for
nodes that share a common set of parameters (such as all the Buffer* and
AnimatedBuffer* flavours). For example, `BufferUBVec4(name='test')`
wouldn't work before this commit.

As a side effect of this change, on x86 the generated binding size is
reduced from 551K to 447K (about 19% smaller) due to the disappearance
of a bunch of unecessary _init_params() functions.